requirements.txt
mlx>=0.31.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
pytorch-fid>=0.3.0
